+++ web/Scripts/scripts.js
@@ -122,12 +122,54 @@ $(function() {
     });
 
     $(".search-box input").bind("change paste keyup", function () {
-        $(this).val() ? $(".search-box .search-clear").show() : $(".search-box .search-clear").hide();
+        $(this).val() ? $(this).parent().children(".search-clear").show() : $(this).parent().children(".search-clear").hide();
     });
 
     $(".search-box .search-clear").click(function () {
-        $(".search-box input").val("");
+        $(this).parent().children("input").val("");
         $(this).hide();
+    });
+
+    $("#doc-builder-search-box input").bind("change paste keyup", function () {
+        let valExs = !!$(this).val(); 
+        if (valExs) {
+            $(".builder-search-results").empty();
+            for (let method of methodNames) {
+                if (method.toLowerCase().includes($(this).val().toLowerCase())) {
+                    let elem = document.createElement("li");
+                    elem.innerHTML = method;
+                    $(".builder-search-results").append(elem)
+                }
+            }
+        }
+        $(".builder-search-results")[0].innerHTML != "" ? $(".builder-search-results").show() : $(".builder-search-results").hide();
+    })
+
+    $(document).bind("mouseup", function(e) {
+        if (!$(e.target).closest(".builder-search-results").length) {
+            $(".builder-search-results").hide();
+        } else if ($(e.target).closest(".builder-search-results")) {
+            let liText = e.target.closest("li").innerText;
+
+            $(".builder-search-results").hide();
+            $("#doc-builder-search-box input").val(liText);
+            liText = liText.substring(liText.indexOf("— ") + 2);
+            let postParams = liText.split(".");
+
+            let xhr = new XMLHttpRequest();
+            xhr.open("POST", "");
+            xhr.setRequestHeader( 'Content-Type', 'application/json');
+            xhr.send(JSON.stringify( {module: editorType, section: postParams[0].toLowerCase(), method: postParams[1].toLowerCase()}));
+
+            xhr.onload = function () {
+                let script = xhr.responseText;
+                script = "\n\n" + script.substring(script.indexOf(";") + 2, script.length)
+
+                $("#builderScript").val(function () {
+                    return this.value.substring(0, this.value.indexOf("builder.SaveFile") - 1) + script;
+                });
+            }
+        }
     });
 
     $(".side-nav").treeview({